backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy[asyncio]>=2.0.23
python-multipart>=0.0.6
requests>=2.31.0
numpy>=1.26.0
pandas>=2.2.3
scikit-learn>=1.6.1
python-dateutil==2.8.2
aiofiles==23.2.1
SpeechRecognition>=3.10.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-dotenv>=1.0.0
pgvector>=0.3.0
groq>=0.4.0
gTTS>=2.3.0
xgboost>=2.0.0
langgraph>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.0.4
 